externalise LabsFeature
