+++ app/lib/features/settings/providers/settings_providers.dart
@@ -1,43 +1,12 @@
-import 'dart:convert';
-
 import 'package:acter/common/providers/common_providers.dart';
-import 'package:acter/common/utils/feature_flagger.dart';
 import 'package:acter/common/utils/main.dart';
-import 'package:acter/common/utils/utils.dart';
 import 'package:acter/features/home/providers/client_providers.dart';
-import 'package:acter/features/settings/providers/notifiers/labs_features.dart';
 import 'package:acter/features/settings/providers/notifiers/locale_notifier.dart';
-import 'package:acter_flutter_sdk/acter_flutter_sdk.dart';
 import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 final allowSentryReportingProvider =
     FutureProvider((ref) => getCanReportToSentry());
-
-const labsKey = 'a3.labs';
-
-final asyncFeaturesProvider =
-    FutureProvider<Features<LabsFeature>>((ref) async {
-  final prefInstance = await sharedPrefs();
-  final currentData = prefInstance.getString(labsKey) ?? '[]';
-  final features = featureFlagsFromJson<LabsFeature>(
-    json.decode(currentData),
-    (name) => LabsFeature.values.byName(name),
-  );
-  return Features<LabsFeature>(
-    flags: features,
-    defaultOn: LabsFeature.defaults,
-  );
-});
-
-final featuresProvider =
-    StateNotifierProvider<SharedPrefFeaturesNotifier, Features<LabsFeature>>(
-        (ref) {
-  return SharedPrefFeaturesNotifier(
-    labsKey,
-    ref,
-  );
-});
 
 final localeProvider =
     StateNotifierProvider<LocaleNotifier, String>((ref) => LocaleNotifier());
@@ -75,22 +44,3 @@ final possibleEmailToAddForPushProvider =
   }
   return allowedEmails;
 });
-
-final isActiveProvider = StateProvider.family<bool, LabsFeature>(
-  (ref, feature) => ref.watch(featuresProvider).isActive(feature),
-);
-
-final asyncIsActiveProvider =
-    FutureProvider.family<bool, LabsFeature>((ref, feature) async {
-  return (await ref.watch(asyncFeaturesProvider.future)).isActive(feature);
-});
-
-// helper
-Future<bool> updateFeatureState(
-  WidgetRef ref,
-  LabsFeature f,
-  bool value,
-) async {
-  await ref.read(featuresProvider.notifier).setActive(f, value);
-  return ref.read(featuresProvider).isActive(f);
-}
