+++ app/lib/router/shell_routers/chat_shell_router.dart
@@ -6,7 +6,8 @@ import 'package:acter/features/chat/pages/room_profile_page.dart';
 import 'package:acter/features/chat/widgets/chat_layout_builder.dart';
 import 'package:acter/features/chat_ng/widgets/rooms_list.dart';
 import 'package:acter/features/invite_members/pages/invite_page.dart';
-import 'package:acter/features/settings/providers/settings_providers.dart';
+import 'package:acter/features/labs/model/labs_features.dart';
+import 'package:acter/features/labs/providers/labs_providers.dart';
 import 'package:acter/features/space/settings/pages/visibility_accessibility_page.dart';
 import 'package:acter/router/router.dart';
 import 'package:flutter/material.dart';
